requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0